Create docker environment
The multi stage Dockerfile is made so that (when using docker buildkit), the gradle dependencies will be cached. There's one image to build the jar, another to run it so the runtime image stays clean.

The docker-compose environment gives us access to Elasticsearch in a specifically named network so that later - for instance in a Github Action build - it's easy to connect a container to that network.