Setup dependencies and quality checks
We use the most recent versions of everything, except Elasticsearch since (for now) I want to make this version compatible with ES 6.8.

The operator framework is the base of it all, on top of that we use the spring boot starter so we get Spring Boot's features things like DI, reading environment variables and what not.

We add spotless with ktlint to make sure everyone sticks to the Kotlin code style.