Create resource handler
Since the actions on templates and indices are all executed in a generic way, with the same error handling, we bundle that in this resource handler.

We're also creating an interface with a default implementation so that it's easy to create a client from anything that has a hosts property defined, which our custom resources implementations will have.