Add template resource and controller
The first actual gist that this operator will handle: templates!

In both the CRD schema and the class, we leave the "defnition" field where the template is defined vague - an object and map respectively. By doing this it doesn't matter how future definitions of a template in ES look; our operator will not have to change. It does however make use of the validation of the ES library (caught in the ResourceHandler) to validate the definition before it's sent to the cluster.
To not bind this operator to a specific cluster, we let the user define the hosts on the CRD itself. This way the operator can be deployed once but operate on several clusters at the same time.

The result of the operation will be written back in the status object of the CRD; even if it fails. The status subresource is probably a better way to solve this, leaving a history of failed and successful deploys. Something for a later day.

We make use of MDC to clarify in the logs which CRD we're currently working on.