Add index resource and controller
And the second gist: indices!

This one's a little more complicated than the templates one. That's mainly because updating an index looks different than creating one.

First: each part of the index needs to be updated independently - settings, mappings and aliases - instead of putting in in a single PUT action.
Second: not all index settings can be updated so we have to only use dynamic setting in the update.
Third: aliases have to be independently added or removed

The approach with hosts and the actual index definition is the same as with the templates, see the previous commit for an explanation.